+++ cjdb/modules/importer.py
@@ -14,8 +14,7 @@ from cjdb.logger import logger
 from cjdb.model.sqlalchemy_models import (BaseModel,
                                           CityObjectRelationshipModel,
                                           CjMetadataModel, CjObjectModel)
-from cjdb.modules.checks import (check_object_type,
-                                 check_root_properties)
+from cjdb.modules.checks import check_object_type, check_root_properties
 from cjdb.modules.extensions import ExtensionHandler
 from cjdb.modules.geometric import (get_ground_geometry, get_srid,
                                     reproject_vertex_list,
@@ -41,10 +40,19 @@ class SingleFileImport:
 
 # importer class called once per whole import
 class Importer:
-    def __init__(self, engine, filepath, db_schema, input_srid,
-                 indexed_attributes, partial_indexed_attributes,
-                 ignore_repeated_file, overwrite, transform,
-                 clustering):
+    def __init__(
+        self,
+        engine,
+        filepath,
+        db_schema,
+        input_srid,
+        indexed_attributes,
+        partial_indexed_attributes,
+        ignore_repeated_file,
+        overwrite,
+        transform,
+        clustering,
+    ):
         self.engine = engine
         self.filepath = filepath
         self.db_schema = db_schema
@@ -101,7 +109,7 @@ class Importer:
         # create all tables defined as SqlAlchemy models
         for table in BaseModel.metadata.tables.values():
             table.create(self.engine, checkfirst=True)
-        
+
         # create indexes
         self.create_indexes()
 
@@ -134,13 +142,14 @@ class Importer:
 
     def post_import(self) -> None:
         """Perform post import operation on the schema,
-           like clustering and indexing"""
-  
+        like clustering and indexing"""
+
         if self.clustering:
             self.cluster_tables()
         self.index_attributes()
 
         # Optionally, you can define the cluster_tables method if you want to keep it for future use
+
     def cluster_tables(self) -> None:
         """Cluster tables to improve query performance."""
         logger.info("Clustering tables, this will take some time...")
@@ -163,11 +172,12 @@ class Importer:
         if not self.transform:
             self.current.target_srid = self.current.source_srid
         else:
-            schema_srid = (self.session.query(CjMetadataModel)
-                           .filter(CjMetadataModel.finished_at.isnot(None))
-                           .order_by(CjMetadataModel.finished_at.desc())
-                           .first()
-                           )
+            schema_srid = (
+                self.session.query(CjMetadataModel)
+                .filter(CjMetadataModel.finished_at.isnot(None))
+                .order_by(CjMetadataModel.finished_at.desc())
+                .first()
+            )
             if schema_srid:
                 self.current.target_srid = schema_srid.srid
             else:
@@ -189,13 +199,15 @@ class Importer:
 
         if self.input_srid:
             if self.current.source_srid:
-                logger.warning("""Input SRID is different than the SRID in the 
+                logger.warning(
+                    """Input SRID is different than the SRID in the 
                                 file's metadata:
                                 Input SRID: %s
                                 Source SRID: %s
                                 Source SRID will be overwritten.""",
-                               self.input_srid,
-                               self.current.source_srid)
+                    self.input_srid,
+                    self.current.source_srid,
+                )
 
             self.current.source_srid = self.input_srid
         else:
@@ -207,8 +219,8 @@ class Importer:
     def extract_cj_metadatadata(self, line_json):
         if "metadata" not in line_json:
             raise exceptions.InvalidMetadataException(
-                "The file should contain a member"
-                "'metadata', in the first object")
+                "The file should contain a member'metadata', in the first object"
+            )
 
         extra_root_properties = find_extra_properties(line_json)
 
@@ -240,9 +252,7 @@ class Importer:
             )
 
         # store extensions data - extra root properties, extra city objects
-        self.current.extension_handler = ExtensionHandler(
-            line_json.get("extensions")
-        )
+        self.current.extension_handler = ExtensionHandler(line_json.get("extensions"))
 
         # prepare extra properties coming from extensions
         # they will be placed in the extra_properties jsonb column
@@ -272,14 +282,13 @@ class Importer:
 
         if cj_metadata.source_file.lower() != "stdin":
             # compare to existing import metas
-            imported_files = cj_metadata.get_already_imported_files(
-                self.session
-            )
+            imported_files = cj_metadata.get_already_imported_files(self.session)
 
             if (
-                self.ignore_repeated_file and
-                    imported_files.first() and 
-                    not self.overwrite):
+                self.ignore_repeated_file
+                and imported_files.first()
+                and not self.overwrite
+            ):
                 logger.warning("File already imported. Skipping...")
                 return False
             elif imported_files.first() and not self.overwrite:
@@ -287,7 +296,8 @@ class Importer:
                     "A file with the same name (%s) was previously "
                     "imported on %s. Use the --ignore-repeated-file "
                     "to skip already imported files.",
-                    cj_metadata.source_file, imported_files.first().finished_at
+                    cj_metadata.source_file,
+                    imported_files.first().finished_at,
                 )
                 user_answer = input(
                     "Should the import continue? "
@@ -298,21 +308,25 @@ class Importer:
                 )
                 if user_answer.lower() != "y":
                     logger.warning(
-                        f"Import of file {cj_metadata.source_file}"
-                        "skipped by user.")
+                        f"Import of file {cj_metadata.source_file}skipped by user."
+                    )
                     return False
             elif imported_files.first() and self.overwrite:
                 logger.warning(
                     "File already imported. Overwriting all objects"
-                    f" from source file {cj_metadata.source_file}")
+                    f" from source file {cj_metadata.source_file}"
+                )
                 imported_files.delete()
 
         different_srid = cj_metadata.different_srid_meta(self.session)
         if different_srid:
-            logger.error("Not matching coordinate Reference Systems"
-                         "\nCurrently imported SRID: %s"
-                         "\nRecently imported SRID: %s",
-                         cj_metadata.srid, different_srid.srid)
+            logger.error(
+                "Not matching coordinate Reference Systems"
+                "\nCurrently imported SRID: %s"
+                "\nRecently imported SRID: %s",
+                cj_metadata.srid,
+                different_srid.srid,
+            )
             raise exceptions.InconsistentCRSException()
 
         # add metadata to the database
@@ -333,9 +347,7 @@ class Importer:
 
         # reproject if needed
         source_target_srid = None
-        if (
-            self.current.target_srid != self.current.source_srid
-        ):
+        if self.current.target_srid != self.current.source_srid:
             source_target_srid = (
                 self.current.source_srid,
                 self.current.target_srid,
@@ -349,7 +361,6 @@ class Importer:
 
         # create CityJSONObjects
         for obj_id, cityobj in line_json["CityObjects"].items():
-
             # get 3D geom, ground geom and bbox
             geometry, ground_geometry = self.get_geometries(
                 obj_id, cityobj, vertices, source_target_srid
@@ -393,19 +404,18 @@ class Importer:
             for child_id in cityobj.get("children", []):
                 child_unique_id = self.processed.get(child_id, None)
                 if child_unique_id:
-                    city_object_relationships_ties.append((city_object_id,
-                                                           child_unique_id))
+                    city_object_relationships_ties.append(
+                        (city_object_id, child_unique_id)
+                    )
                 else:
                     self.max_id = self.max_id + 1
                     self.processed[child_id] = self.max_id
-                    city_object_relationships_ties.append((city_object_id,
-                                                           self.max_id))
+                    city_object_relationships_ties.append((city_object_id, self.max_id))
 
         # create children-parent links after all objects
         # from the CityJSONFeature already exist
         for parent_id, child_id in city_object_relationships_ties:
-            self.current.families.append({"parent_id": parent_id,
-                                          "child_id": child_id})
+            self.current.families.append({"parent_id": parent_id, "child_id": child_id})
 
     def process_file(self, filepath) -> bool:
         """Process a single cityJSON file"""
@@ -429,7 +439,6 @@ class Importer:
         for line in f.readlines():
             line_json = json.loads(line.rstrip("\n"))
             self.process_line(line_json)
-        logger.debug(f"Lines to read {len(self.current.city_objects)}")
         if self.current.city_objects:
             logger.debug("Importing city objects")
             obj_insert = (
@@ -481,9 +490,7 @@ class Importer:
         )
 
         # prepare partial and non partial indexes in one list
-        attributes = [
-            (a, True) for a in self.partial_indexed_attributes
-        ] + [  # noqa
+        attributes = [(a, True) for a in self.partial_indexed_attributes] + [  # noqa
             (a, False) for a in self.indexed_attributes
         ]